+++ src/sorters/CocktailSort.java
@@ -7,7 +7,7 @@ public class CocktailSort extends Sorter {
     }
 
     @Override
-    public void sort(int delay, boolean sortAscending) {
+    public void sort(double delay, boolean sortAscending) {
         shouldStop = false;
         boolean sorted = false;
         while (!sorted) {
